debug select2 value after validation fails

+++ resources/views/components/milestone/fields.blade.php
@@ -15,7 +15,7 @@
                     <select class="form-control @error('owner_id') is-invalid @enderror" name="owner_id" id="owner-id" style="width: 100%;">
                         <option disabled selected value>select owner</option>
                         @foreach($project->team as $user)
-                            <option value="{{ $user->id }}" @selected(old('owner_id', $milestone->owner_id ?? null) === $user->id)>{{ $user->full_name }}</option>
+                            <option value="{{ $user->id }}" @selected((int) old('owner_id', $milestone->owner_id ?? null) === $user->id)>{{ $user->full_name }}</option>
                         @endforeach
                     </select>
                     @error('owner_id')
